+++ Assets/NavVolume/Runtime/Pathfinding/SVOPathfinder.cs
@@ -238,7 +238,7 @@ namespace NavVolume.Runtime.Pathfinding
             }
             else if (link.IsVoxel(out var subnodeIdx))
             {
-                var (x, y, z) = SVOLeaf.IndexToSubnodeCoords((int)subnodeIdx);
+                var (x, y, z) = SVOLeaf.IndexToSubnodeCoords(subnodeIdx);
                 TryAddVoxelNeighbor(svo, link, x + 1, y, z, NeighborDirection.PosX, buffer);
                 TryAddVoxelNeighbor(svo, link, x - 1, y, z, NeighborDirection.NegX, buffer);
                 TryAddVoxelNeighbor(svo, link, x, y + 1, z, NeighborDirection.PosY, buffer);
@@ -273,7 +273,7 @@ namespace NavVolume.Runtime.Pathfinding
             )
             {
                 var subnode = SVOLeaf.SubnodeCoordsToIndex(nx, ny, nz);
-                buffer.Add(SVOLink.VoxelLink(voxelLink.Offset, (uint)subnode));
+                buffer.Add(SVOLink.VoxelLink(voxelLink.Offset, subnode));
                 return;
             }
 
@@ -296,7 +296,7 @@ namespace NavVolume.Runtime.Pathfinding
             if (layerIdx == 0)
             {
                 var entrySubnode = SVOLeaf.SubnodeCoordsToIndex(nx, ny, nz);
-                buffer.Add(SVOLink.VoxelLink(leafNeighbor.Offset, (uint)entrySubnode));
+                buffer.Add(SVOLink.VoxelLink(leafNeighbor.Offset, entrySubnode));
             }
             else
             {
@@ -343,7 +343,7 @@ namespace NavVolume.Runtime.Pathfinding
             var childLayer = coarseNeighborLIdx - 1;
             var firstOffset = coarse.FirstChild.Offset;
 
-            for (var c = 0u; c < 8; c++)
+            for (var c = 0; c < 8; c++)
             {
                 var childLink = SVOLink.NodeLink(childLayer, firstOffset + c);
 
@@ -425,7 +425,7 @@ namespace NavVolume.Runtime.Pathfinding
                     continue;
                 }
 
-                var voxelLink = SVOLink.VoxelLink(leafLink.Offset, (uint)i);
+                var voxelLink = SVOLink.VoxelLink(leafLink.Offset, i);
                 if (_closed.Contains(voxelLink))
                 {
                     continue;
@@ -496,7 +496,7 @@ namespace NavVolume.Runtime.Pathfinding
 
             var leaf = svo.LeafNodes[link.Offset];
             link.IsVoxel(out var subnodeIdx);
-            return leaf.IsOccupied((int)subnodeIdx);
+            return leaf.IsOccupied(subnodeIdx);
         }
 
         PathResult ReconstructPath(
